Add abstract controller
