+++ Classes/Controller/PasswordController.php
@@ -13,27 +13,14 @@ namespace PAGEmachine\Hairu\Controller;
  * LICENSE.txt file that was distributed with this source code.
  */
 
-use PAGEmachine\Hairu\Mvc\Controller\ActionController;
 use TYPO3\CMS\Core\Messaging\FlashMessage;
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 
 /**
  * Controller for password tasks
  */
-class PasswordController extends ActionController
+class PasswordController extends AbstractController
 {
-    /**
-     * @var \PAGEmachine\Hairu\Domain\Repository\FrontendUserRepository
-     * @inject
-     */
-    protected $frontendUserRepository;
-
-    /**
-     * @var \PAGEmachine\Hairu\Domain\Service\AuthenticationService
-     * @inject
-     */
-    protected $authenticationService;
-
     /**
      * Update password form view
      *
@@ -87,26 +74,5 @@ class PasswordController extends ActionController
 
             $this->forward('showPasswordEditForm');
         }
-    }
-
-    /**
-     * Shorthand helper for adding localized flash messages
-     *
-     * @param string $translationKey
-     * @param array $translationArguments
-     * @param int $severity
-     * @param string $messageTitle
-     */
-    protected function addLocalizedFlashMessage($translationKey, array $translationArguments = null, $severity = FlashMessage::OK, $messageTitle = '')
-    {
-        $this->addFlashMessage(
-            LocalizationUtility::translate(
-                $translationKey,
-                $this->request->getControllerExtensionName(),
-                $translationArguments
-            ),
-            ($messageTitle != '' ? LocalizationUtility::translate($messageTitle, $this->request->getControllerExtensionName(), $translationArguments) : ''),
-            $severity
-        );
     }
 }
